+++ src/app/modules/openlayers.module.ts
@@ -32,3 +32,4 @@ export { default as WKT } from 'ol/format/WKT';
 export { getCenter } from 'ol/extent';
 export { getDistance } from 'ol/sphere';
 export { createXYZ } from 'ol/tilegrid';
+